Move InitialSearchFilters declaration
