+++ packages/web/components/Dashboard/Filters/index.ts
@@ -1,2 +1,2 @@
 export { default } from './Filters'
-export type { PostQueryVarsType, InitialSearchFilters } from './Filters'
+export type { PostQueryVarsType } from './Filters'
