+++ packages/web/components/Dashboard/MyFeed/index.ts
@@ -1,1 +1,2 @@
 export { default } from './MyFeed'
+export type { InitialSearchFilters } from './MyFeed'
