+++ packages/web/components/InlineFeedbackPopover/Comment.tsx
@@ -5,8 +5,6 @@ import classNames from 'classnames'
 import Markdown from 'react-markdown'
 import remarkGfm from 'remark-gfm'
 
-// TODO: Update `react-markdown` and `remark-gfm` once Next.js versioning issues are resolved
-
 import {
   useUpdateCommentMutation,
   useDeleteCommentMutation,
